Do not fail the delete if providerConfig is not found
When buildup of a cluster fails the state of the object can be
in different unfinished levels. If some required arguments are missing,
nothing is build up. The object can also not be deleted because early
stages fail.
Accept missing providerConfig as a non build up cluster in delete operations.